+++ packages/editor/src/lib/editor/Editor.ts
@@ -131,8 +131,8 @@ import { TLFrameShape } from './shapes/FrameShape/frameShapeTypes'
 import { GroupShapeUtil } from './shapes/GroupShape/GroupShapeUtil/GroupShapeUtil'
 import { TLGroupShape } from './shapes/GroupShape/groupShapeTypes'
 import { TextShapeUtil } from './shapes/TextShape/TextShapeUtil/TextShapeUtil'
+import { TLExportColors } from './shapes/shared/TLExportColors'
 import { ShapeUtil, TLResizeMode } from './shapeutils/ShapeUtil'
-import { TLExportColors } from './shapeutils/shared/TLExportColors'
 import { RootState } from './tools/RootState'
 import { StateNode, TLStateNodeConstructor } from './tools/StateNode'
 import { TLContent } from './types/clipboard-types'
