+++ packages/editor/src/lib/editor/managers/TextManager.ts
@@ -1,7 +1,7 @@
 import { Box2dModel, TLAlignType } from '@tldraw/tlschema'
 import { uniqueId } from '../../utils/data'
 import { Editor } from '../Editor'
-import { TextHelpers } from '../shapeutils/shared/TextHelpers'
+import { TextHelpers } from '../shapes/shared/TextHelpers'
 
 const textAlignmentsForLtr = {
 	start: 'left',
