+++ packages/editor/src/lib/editor/shapes/DrawShape/DrawShapeUtil/DrawShapeUtil.tsx
@@ -15,9 +15,9 @@ import { last, rng } from '@tldraw/utils'
 import { SVGContainer } from '../../../../components/SVGContainer'
 import { getSvgPathFromStroke, getSvgPathFromStrokePoints } from '../../../../utils/svg'
 import { ShapeUtil, TLOnResizeHandler } from '../../../shapeutils/ShapeUtil'
-import { ShapeFill, getShapeFillSvg } from '../../../shapeutils/shared/ShapeFill'
-import { TLExportColors } from '../../../shapeutils/shared/TLExportColors'
-import { useForceSolid } from '../../../shapeutils/shared/useForceSolid'
+import { ShapeFill, getShapeFillSvg } from '../../shared/ShapeFill'
+import { TLExportColors } from '../../shared/TLExportColors'
+import { useForceSolid } from '../../shared/useForceSolid'
 import { TLDrawShape, TLDrawShapeSegment } from '../drawShapeTypes'
 import { getDrawShapeStrokeDashArray, getFreehandOptions, getPointsFromSegments } from './getPath'
