+++ packages/editor/src/lib/editor/shapes/FrameShape/FrameShapeUtil/FrameShapeUtil.tsx
@@ -5,8 +5,8 @@ import { SVGContainer } from '../../../../components/SVGContainer'
 import { defaultEmptyAs } from '../../../../utils/string'
 import { BaseBoxShapeUtil } from '../../../shapeutils/BaseBoxShapeUtil'
 import { TLOnResizeEndHandler } from '../../../shapeutils/ShapeUtil'
-import { createTextSvgElementFromSpans } from '../../../shapeutils/shared/createTextSvgElementFromSpans'
-import { TLExportColors } from '../../../shapeutils/shared/TLExportColors'
+import { createTextSvgElementFromSpans } from '../../shared/createTextSvgElementFromSpans'
+import { TLExportColors } from '../../shared/TLExportColors'
 import { TLFrameShape } from '../frameShapeTypes'
 import { FrameHeading } from './components/FrameHeading'
