+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeUtil/GeoShapeUtil.tsx
@@ -18,11 +18,11 @@ import { FONT_FAMILIES, LABEL_FONT_SIZES, TEXT_PROPS } from '../../../../constan
 import { Editor } from '../../../Editor'
 import { BaseBoxShapeUtil } from '../../../shapeutils/BaseBoxShapeUtil'
 import { TLOnEditEndHandler, TLOnResizeHandler } from '../../../shapeutils/ShapeUtil'
-import { getTextLabelSvgElement } from '../../../shapeutils/shared/getTextLabelSvgElement'
-import { HyperlinkButton } from '../../../shapeutils/shared/HyperlinkButton'
-import { TextLabel } from '../../../shapeutils/shared/TextLabel'
-import { TLExportColors } from '../../../shapeutils/shared/TLExportColors'
-import { useForceSolid } from '../../../shapeutils/shared/useForceSolid'
+import { getTextLabelSvgElement } from '../../shared/getTextLabelSvgElement'
+import { HyperlinkButton } from '../../shared/HyperlinkButton'
+import { TextLabel } from '../../shared/TextLabel'
+import { TLExportColors } from '../../shared/TLExportColors'
+import { useForceSolid } from '../../shared/useForceSolid'
 import { TLGeoShape } from '../geoShapeTypes'
 import { DashStyleEllipse, DashStyleEllipseSvg } from './components/DashStyleEllipse'
 import { DashStyleOval, DashStyleOvalSvg } from './components/DashStyleOval'
