+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeUtil/components/DashStyleOval.tsx
@@ -1,13 +1,9 @@
 import { toDomPrecision } from '@tldraw/primitives'
 import { TLShapeId } from '@tldraw/tlschema'
 import React from 'react'
-import {
-	ShapeFill,
-	getShapeFillSvg,
-	getSvgWithShapeFill,
-} from '../../../../shapeutils/shared/ShapeFill'
-import { TLExportColors } from '../../../../shapeutils/shared/TLExportColors'
-import { getPerfectDashProps } from '../../../../shapeutils/shared/getPerfectDashProps'
+import { ShapeFill, getShapeFillSvg, getSvgWithShapeFill } from '../../../shared/ShapeFill'
+import { TLExportColors } from '../../../shared/TLExportColors'
+import { getPerfectDashProps } from '../../../shared/getPerfectDashProps'
 import { TLGeoShape } from '../../geoShapeTypes'
 import { getOvalPerimeter, getOvalSolidPath } from '../helpers'
