+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeUtil/components/DashStylePolygon.tsx
@@ -1,12 +1,8 @@
 import { Vec2d, VecLike } from '@tldraw/primitives'
 import React from 'react'
-import {
-	ShapeFill,
-	getShapeFillSvg,
-	getSvgWithShapeFill,
-} from '../../../../shapeutils/shared/ShapeFill'
-import { TLExportColors } from '../../../../shapeutils/shared/TLExportColors'
-import { getPerfectDashProps } from '../../../../shapeutils/shared/getPerfectDashProps'
+import { ShapeFill, getShapeFillSvg, getSvgWithShapeFill } from '../../../shared/ShapeFill'
+import { TLExportColors } from '../../../shared/TLExportColors'
+import { getPerfectDashProps } from '../../../shared/getPerfectDashProps'
 import { TLGeoShape } from '../../geoShapeTypes'
 
 export const DashStylePolygon = React.memo(function DashStylePolygon({
