+++ packages/editor/src/lib/editor/shapes/GeoShape/GeoShapeUtil/components/SolidStyleEllipse.tsx
@@ -1,10 +1,6 @@
 import React from 'react'
-import {
-	ShapeFill,
-	getShapeFillSvg,
-	getSvgWithShapeFill,
-} from '../../../../shapeutils/shared/ShapeFill'
-import { TLExportColors } from '../../../../shapeutils/shared/TLExportColors'
+import { ShapeFill, getShapeFillSvg, getSvgWithShapeFill } from '../../../shared/ShapeFill'
+import { TLExportColors } from '../../../shared/TLExportColors'
 import { TLGeoShape } from '../../geoShapeTypes'
 
 export const SolidStyleEllipse = React.memo(function SolidStyleEllipse({
