+++ packages/editor/src/lib/editor/shapes/NoteShape/NoteShapeUtil/NoteShapeUtil.tsx
@@ -2,10 +2,10 @@ import { Box2d, Vec2d, toDomPrecision } from '@tldraw/primitives'
 import { FONT_FAMILIES, LABEL_FONT_SIZES, TEXT_PROPS } from '../../../../constants'
 import { Editor } from '../../../Editor'
 import { ShapeUtil, TLOnEditEndHandler } from '../../../shapeutils/ShapeUtil'
-import { HyperlinkButton } from '../../../shapeutils/shared/HyperlinkButton'
-import { TLExportColors } from '../../../shapeutils/shared/TLExportColors'
-import { TextLabel } from '../../../shapeutils/shared/TextLabel'
-import { getTextLabelSvgElement } from '../../../shapeutils/shared/getTextLabelSvgElement'
+import { HyperlinkButton } from '../../shared/HyperlinkButton'
+import { TLExportColors } from '../../shared/TLExportColors'
+import { TextLabel } from '../../shared/TextLabel'
+import { getTextLabelSvgElement } from '../../shared/getTextLabelSvgElement'
 import { TLNoteShape } from '../noteShapeTypes'
 
 const NOTE_SIZE = 200
