+++ packages/editor/src/lib/editor/shapes/shared/DashedOutlineBox.tsx
