+++ packages/editor/src/lib/editor/shapes/shared/TextHelpers.ts
