+++ packages/editor/src/lib/editor/shapes/shared/createTextSvgElementFromSpans.ts
