+++ packages/editor/src/lib/editor/shapes/shared/getDrawStrokeInfo.ts
