+++ packages/editor/src/lib/editor/shapes/shared/getPerfectDashProps.ts
