+++ packages/editor/src/lib/editor/shapes/shared/getStrokeDashArray.ts
