+++ packages/editor/src/lib/editor/shapes/shared/getTextLabelSvgElement.ts
