+++ packages/editor/src/lib/editor/shapes/shared/isBlackColor.ts
