+++ packages/editor/src/lib/editor/shapes/shared/resizeScaled.ts
