+++ packages/editor/src/lib/editor/shapes/shared/useEditableText.ts
